# apps/api/openapi/openapi.json
{
  "openapi": "3.1.0",
...
  "info": {
    "title": "Sadat Real Estate API",
    "version": "0.0.0",
    "description": "Executable contract for implemented routes only. Product routes use /api/v1; operational probes remain at the process root."
  },
  "servers": [
    {
      "url": "{baseUrl}",
      "description": "Local synthetic default; replace with an explicitly approved non-production or production origin.",
      "variables": {
        "baseUrl": {
          "default": "http://127.0.0.1:3000"
        }
      }
    }
  ],
...
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "operationId": "loginAdmin",
        "summary": "Authenticate an Admin with normalized email and password",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AdminLoginRequest" }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/AuthSessionSuccess" },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "401": { "$ref": "#/components/responses/AuthError" },
          "403": { "$ref": "#/components/responses/AuthError" },
          "413": { "$ref": "#/components/responses/AuthError" },
          "429": { "$ref": "#/components/responses/AuthError" },
          "500": { "$ref": "#/components/responses/AuthError" }
        }
      }
    },
    "/api/v1/auth/otp/send": {
      "post": {
        "operationId": "sendPhoneOtp",
        "summary": "Create and deliver a bounded phone OTP challenge",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/OtpSendRequest" }
            }
          }
        },
        "responses": {
          "202": { "$ref": "#/components/responses/OtpSendSuccess" },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "413": { "$ref": "#/components/responses/AuthError" },
          "429": { "$ref": "#/components/responses/AuthError" },
          "500": { "$ref": "#/components/responses/AuthError" },
          "503": { "$ref": "#/components/responses/AuthError" }
        }
      }
    },
    "/api/v1/auth/otp/verify": {
      "post": {
        "operationId": "verifyPhoneOtp",
        "summary": "Verify a phone OTP for login or a one-time registration authority",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/OtpVerifyRequest" }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/OtpVerifySuccess" },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "401": { "$ref": "#/components/responses/AuthError" },
          "403": { "$ref": "#/components/responses/AuthError" },
          "413": { "$ref": "#/components/responses/AuthError" },
          "429": { "$ref": "#/components/responses/AuthError" },
          "500": { "$ref": "#/components/responses/AuthError" }
        }
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "operationId": "refreshSession",
        "summary": "Rotate an opaque refresh session",
        "security": [{ "refreshCookie": [] }],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/EmptyAuthRequest" } }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/AuthSessionSuccess" },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "401": { "$ref": "#/components/responses/AuthError" },
          "403": { "$ref": "#/components/responses/AuthError" },
          "413": { "$ref": "#/components/responses/AuthError" },
          "429": { "$ref": "#/components/responses/AuthError" },
          "500": { "$ref": "#/components/responses/AuthError" }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "logoutSession",
        "summary": "Revoke the current refresh session",
        "security": [{ "refreshCookie": [] }],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/EmptyAuthRequest" } }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/LogoutSuccess" },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "401": { "$ref": "#/components/responses/AuthError" },
          "413": { "$ref": "#/components/responses/AuthError" },
          "429": { "$ref": "#/components/responses/AuthError" },
          "500": { "$ref": "#/components/responses/AuthError" }
        }
      }
    },
    "/api/v1/auth/register/seeker": {
      "post": {
        "operationId": "registerSeeker",
        "summary": "Create a seeker account from a verified phone registration authority",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/SeekerRegistrationRequest" } }
          }
        },
        "responses": {
          "201": { "$ref": "#/components/responses/SeekerRegistrationSuccess" },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "401": { "$ref": "#/components/responses/SeekerError" },
          "409": { "$ref": "#/components/responses/SeekerError" },
          "413": { "$ref": "#/components/responses/SeekerError" },
          "429": { "$ref": "#/components/responses/SeekerError" },
          "500": { "$ref": "#/components/responses/SeekerError" }
        }
      }
    },
    "/api/v1/me": {
      "get": {
        "operationId": "getCurrentSeekerProfile",
        "summary": "Read the authenticated seeker's own profile",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "$ref": "#/components/responses/SeekerProfileSuccess" },
          "401": { "$ref": "#/components/responses/SeekerError" },
          "403": { "$ref": "#/components/responses/SeekerError" },
          "404": { "$ref": "#/components/responses/SeekerError" }
        }
      },
      "patch": {
        "operationId": "updateCurrentSeekerProfile",
        "summary": "Update allowlisted fields on the authenticated seeker's own profile",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/SeekerProfilePatch" } }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/SeekerProfileSuccess" },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "401": { "$ref": "#/components/responses/SeekerError" },
          "403": { "$ref": "#/components/responses/SeekerError" },
          "404": { "$ref": "#/components/responses/SeekerError" }
        }
      }
    },
    "/api/v1/me/preferences": {
      "get": {
        "operationId": "getCurrentSeekerPreferences",
        "summary": "Read the authenticated seeker's own search preferences",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "$ref": "#/components/responses/SeekerPreferencesSuccess" },
          "401": { "$ref": "#/components/responses/SeekerError" },
          "403": { "$ref": "#/components/responses/SeekerError" },
          "404": { "$ref": "#/components/responses/SeekerError" }
        }
      },
      "patch": {
        "operationId": "updateCurrentSeekerPreferences",
        "summary": "Update allowlisted fields on the authenticated seeker's own search preferences",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/SeekerPreferencesPatch" } }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/SeekerPreferencesSuccess" },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "401": { "$ref": "#/components/responses/SeekerError" },
          "403": { "$ref": "#/components/responses/SeekerError" },
          "404": { "$ref": "#/components/responses/SeekerError" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "refreshCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sadat_refresh",
        "description": "Opaque rotating refresh token. Never exposed in a response body."
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Short-lived access token issued by the shared authentication session boundary."
      }
    },
    "responses": {
      "AuthSessionSuccess": {
        "description": "Authenticated session established; refresh token is set as an HttpOnly cookie",
        "headers": {
          "X-Request-Id": { "$ref": "#/components/headers/X-Request-Id" },
          "Traceparent": { "$ref": "#/components/headers/Traceparent" },
          "Set-Cookie": {
            "description": "Rotating sadat_refresh cookie with HttpOnly, SameSite=Strict, scoped Path, and Secure outside Local/Test",
            "schema": { "type": "string" }
          }
        },
        "content": {
          "application/json": { "schema": { "$ref": "#/components/schemas/AuthSessionSuccessEnvelope" } }
        }
      },
      "OtpSendSuccess": {
        "description": "OTP challenge accepted for delivery; the code is never returned",
        "headers": {
          "X-Request-Id": { "$ref": "#/components/headers/X-Request-Id" },
          "Traceparent": { "$ref": "#/components/headers/Traceparent" }
        },
        "content": {
          "application/json": { "schema": { "$ref": "#/components/schemas/OtpSendSuccessEnvelope" } }
        }
      },
      "OtpVerifySuccess": {
        "description": "Existing account authenticated or a single-use registration authority issued",
        "headers": {
          "X-Request-Id": { "$ref": "#/components/headers/X-Request-Id" },
          "Traceparent": { "$ref": "#/components/headers/Traceparent" },
          "Set-Cookie": {
            "description": "Present only for an authenticated outcome; sets the rotating sadat_refresh cookie",
            "schema": { "type": "string" }
          }
        },
        "content": {
          "application/json": { "schema": { "$ref": "#/components/schemas/OtpVerifySuccessEnvelope" } }
        }
      },
      "LogoutSuccess": {
        "description": "Current refresh session revoked and cookie cleared",
        "headers": {
          "X-Request-Id": { "$ref": "#/components/headers/X-Request-Id" },
          "Traceparent": { "$ref": "#/components/headers/Traceparent" },
          "Set-Cookie": { "description": "Clears sadat_refresh", "schema": { "type": "string" } }
        },
        "content": {
          "application/json": { "schema": { "$ref": "#/components/schemas/LogoutSuccessEnvelope" } }
        }
      },
      "ValidationError": {
        "description": "Strict request validation failed",
        "headers": {
          "X-Request-Id": { "$ref": "#/components/headers/X-Request-Id" },
          "Traceparent": { "$ref": "#/components/headers/Traceparent" }
        },
        "content": {
          "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } }
        }
      },
      "AuthError": {
        "description": "Authentication, account-state, replay, or rate-limit failure",
        "headers": {
          "X-Request-Id": { "$ref": "#/components/headers/X-Request-Id" },
          "Traceparent": { "$ref": "#/components/headers/Traceparent" }
        },
        "content": {
          "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } }
        }
      },
      "SeekerRegistrationSuccess": {
        "description": "Seeker account created from a verified OTP authority; refresh token is set as an HttpOnly cookie",
        "headers": {
          "X-Request-Id": { "$ref": "#/components/headers/X-Request-Id" },
          "Traceparent": { "$ref": "#/components/headers/Traceparent" },
          "Set-Cookie": { "description": "Rotating sadat_refresh cookie", "schema": { "type": "string" } }
        },
        "content": {
          "application/json": { "schema": { "$ref": "#/components/schemas/SeekerRegistrationSuccessEnvelope" } }
        }
      },
      "SeekerProfileSuccess": {
        "description": "The authenticated seeker's own public profile projection",
        "headers": {
          "X-Request-Id": { "$ref": "#/components/headers/X-Request-Id" },
          "Traceparent": { "$ref": "#/components/headers/Traceparent" }
        },
        "content": {
          "application/json": { "schema": { "$ref": "#/components/schemas/SeekerProfileSuccessEnvelope" } }
        }
      },
      "SeekerPreferencesSuccess": {
        "description": "The authenticated seeker's own search preferences",
        "headers": {
          "X-Request-Id": { "$ref": "#/components/headers/X-Request-Id" },
          "Traceparent": { "$ref": "#/components/headers/Traceparent" }
        },
        "content": {
          "application/json": { "schema": { "$ref": "#/components/schemas/SeekerPreferencesSuccessEnvelope" } }
        }
      },
      "SeekerError": {
        "description": "Seeker authentication, ownership, registration, or validation failure",
        "headers": {
          "X-Request-Id": { "$ref": "#/components/headers/X-Request-Id" },
          "Traceparent": { "$ref": "#/components/headers/Traceparent" }
        },
        "content": {
          "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } }
        }
      }
    },
    "headers": {
      "X-Request-Id": {
        "description": "Validated request correlation identifier; generated when the inbound value is absent or malformed.",
        "schema": { "type": "string", "minLength": 1, "maxLength": 128, "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$" }
      },
      "Traceparent": {
        "description": "W3C trace context for this API span.",
        "schema": { "type": "string", "pattern": "^00-[0-9a-f]{32}-[0-9a-f]{16}-[0-9a-f]{2}$" }
      }
    },
    "schemas": {
...
      "ApiError": {
        "type": "object",
        "additionalProperties": false,
        "required": ["code", "messageKey", "details", "requestId"],
        "properties": {
          "code": { "type": "string", "pattern": "^[A-Z][A-Z0-9]*(?:_[A-Z0-9]+)*$", "maxLength": 64 },
          "messageKey": { "type": "string", "pattern": "^[a-z][a-zA-Z0-9]*(?:[._-][a-zA-Z0-9]+)*$", "maxLength": 160 },
          "details": { "type": "array", "maxItems": 100, "items": { "$ref": "#/components/schemas/ErrorDetail" } },
          "requestId": { "type": "string", "minLength": 1, "maxLength": 128 }
        }
      },
      "ErrorEnvelope": {
        "type": "object",
        "additionalProperties": false,
        "required": ["error"],
        "properties": { "error": { "$ref": "#/components/schemas/ApiError" } }
      },
      "SuccessEnvelope": {
        "type": "object",
        "additionalProperties": false,
        "required": ["data", "meta"],
        "properties": {
          "data": {},
          "meta": { "$ref": "#/components/schemas/ResponseMeta" }
        }
      },
      "AdminLoginRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["email", "password"],
        "properties": {
          "email": { "type": "string", "format": "email", "minLength": 3, "maxLength": 254 },
          "password": { "type": "string", "minLength": 1, "maxLength": 1024, "writeOnly": true }
        }
      },
      "EmptyAuthRequest": {
        "type": "object",
        "additionalProperties": false,
        "maxProperties": 0
      },
      "NormalizedPhone": {
        "type": "string",
        "pattern": "^\\+[1-9]\\d{7,14}$",
        "maxLength": 16,
        "description": "Canonical E.164 phone number after removing allowed presentation separators or converting a 00 prefix"
      },
      "OtpSendRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["phone", "roleType", "purpose"],
        "properties": {
          "phone": { "$ref": "#/components/schemas/NormalizedPhone" },
          "roleType": { "enum": ["seeker", "provider"] },
          "purpose": { "enum": ["login", "registration"] }
        }
      },
      "OtpVerifyRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["phone", "roleType", "purpose", "challengeId", "code"],
        "properties": {
          "phone": { "$ref": "#/components/schemas/NormalizedPhone" },
          "roleType": { "enum": ["seeker", "provider"] },
          "purpose": { "enum": ["login", "registration"] },
          "challengeId": { "type": "string", "format": "uuid" },
          "code": { "type": "string", "pattern": "^\\d{6}$", "writeOnly": true }
        }
      },
      "RegistrationToken": {
        "type": "string",
        "pattern": "^[A-Za-z0-9_-]{43}$",
        "description": "Single-use opaque authority returned by verified registration OTP."
      },
      "SeekerRegistrationRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["verificationToken", "firstName", "lastName"],
        "properties": {
          "verificationToken": { "$ref": "#/components/schemas/RegistrationToken" },
          "firstName": { "type": "string", "minLength": 1, "maxLength": 80 },
          "lastName": { "type": "string", "minLength": 1, "maxLength": 80 },
          "locale": { "enum": ["ar", "en", "zh-CN"] }
        }
      },
      "SeekerProfilePatch": {
        "type": "object",
        "additionalProperties": false,
        "minProperties": 1,
        "properties": {
          "firstName": { "type": "string", "minLength": 1, "maxLength": 80 },
          "lastName": { "type": "string", "minLength": 1, "maxLength": 80 },
          "locale": { "enum": ["ar", "en", "zh-CN"] }
        }
      },
      "SeekerPreferencesPatch": {
        "type": "object",
        "additionalProperties": false,
        "minProperties": 1,
        "properties": {
          "propertyTypes": { "type": "array", "maxItems": 20, "items": { "type": "string", "minLength": 1, "maxLength": 80 } },
          "locations": { "type": "array", "maxItems": 20, "items": { "type": "string", "minLength": 1, "maxLength": 80 } },
          "purpose": { "enum": ["buy", "rent"] },
          "minPrice": { "type": "integer", "minimum": 0, "maximum": 1000000000000 },
          "maxPrice": { "type": "integer", "minimum": 0, "maximum": 1000000000000 },
          "bedroomsMin": { "type": "integer", "minimum": 0, "maximum": 100 },
          "bedroomsMax": { "type": "integer", "minimum": 0, "maximum": 100 }
        }
      },
      "SeekerProfileData": {
        "type": "object",
        "additionalProperties": false,
        "required": ["id", "roleType", "status", "phone", "firstName", "lastName", "locale"],
        "properties": {
          "id": { "type": "string", "pattern": "^[a-f0-9]{24}$" },
          "roleType": { "const": "seeker" },
          "status": { "enum": ["draft", "unverified", "pending_review", "needs_information", "verified", "rejected", "restricted", "suspended"] },
          "phone": { "$ref": "#/components/schemas/NormalizedPhone" },
          "firstName": { "type": "string", "minLength": 1, "maxLength": 80 },
          "lastName": { "type": "string", "minLength": 1, "maxLength": 80 },
          "locale": { "enum": ["ar", "en", "zh-CN"] }
        }
      },
      "SeekerPreferencesData": {
        "type": "object",
        "additionalProperties": false,
        "required": ["preferences", "updatedAt"],
        "properties": {
          "preferences": { "$ref": "#/components/schemas/SeekerPreferences" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "SeekerRegistrationData": {
        "type": "object",
        "additionalProperties": false,
        "required": ["outcome", "session"],
        "properties": {
          "outcome": { "const": "registered" },
          "session": { "$ref": "#/components/schemas/AuthSessionData" }
        }
      },
      "SeekerRegistrationSuccessEnvelope": {
        "allOf": [
          { "$ref": "#/components/schemas/SuccessEnvelope" },
          { "type": "object", "properties": { "data": { "$ref": "#/components/schemas/SeekerRegistrationData" } } }
        ]
      },
      "SeekerProfileSuccessEnvelope": {
        "allOf": [
          { "$ref": "#/components/schemas/SuccessEnvelope" },
          { "type": "object", "properties": { "data": { "$ref": "#/components/schemas/SeekerProfileData" } } }
        ]
      },
      "SeekerPreferencesSuccessEnvelope": {
        "allOf": [
          { "$ref": "#/components/schemas/SuccessEnvelope" },
          { "type": "object", "properties": { "data": { "$ref": "#/components/schemas/SeekerPreferencesData" } } }
        ]
      },
      "OtpSendData": {
        "type": "object",
        "additionalProperties": false,
        "required": ["accepted", "challengeId", "expiresInSeconds", "retryAfterSeconds"],
        "properties": {
          "accepted": { "const": true },
          "challengeId": { "type": "string", "format": "uuid" },
          "expiresInSeconds": { "type": "integer", "minimum": 1 },
          "retryAfterSeconds": { "type": "integer", "minimum": 1 }
        }
      },
...
      "OtpVerifyData": {
        "oneOf": [
          { "$ref": "#/components/schemas/OtpAuthenticatedData" },
          { "$ref": "#/components/schemas/OtpVerifiedData" }
        ],
        "discriminator": { "propertyName": "outcome" }
      },
      "OtpSendSuccessEnvelope": {
        "allOf": [
          { "$ref": "#/components/schemas/SuccessEnvelope" },
          { "type": "object", "properties": { "data": { "$ref": "#/components/schemas/OtpSendData" } } }
        ]
      },
      "OtpVerifySuccessEnvelope": {
        "allOf": [
          { "$ref": "#/components/schemas/SuccessEnvelope" },
          { "type": "object", "properties": { "data": { "$ref": "#/components/schemas/OtpVerifyData" } } }
        ]
      },
...
      "AuthSessionData": {
        "type": "object",
        "additionalProperties": false,
        "required": ["accessToken", "tokenType", "expiresInSeconds", "user"],
        "properties": {
          "accessToken": { "type": "string", "pattern": "^[A-Za-z0-9_-]+\\.[A-Za-z0-9_-]+\\.[A-Za-z0-9_-]+$" },
          "tokenType": { "const": "Bearer" },
          "expiresInSeconds": { "type": "integer", "minimum": 1 },
          "user": { "$ref": "#/components/schemas/AuthenticatedUser" }
        }
      },
      "AuthSessionSuccessEnvelope": {
        "allOf": [
          { "$ref": "#/components/schemas/SuccessEnvelope" },
          {
            "type": "object",
            "properties": { "data": { "$ref": "#/components/schemas/AuthSessionData" } }
          }
        ]
      },
      "LogoutData": {
        "type": "object",
        "additionalProperties": false,
        "required": ["loggedOut"],
        "properties": { "loggedOut": { "const": true } }
      },
      "LogoutSuccessEnvelope": {
        "allOf": [
          { "$ref": "#/components/schemas/SuccessEnvelope" },
          {
            "type": "object",
            "properties": { "data": { "$ref": "#/components/schemas/LogoutData" } }
          }
        ]
      },
...
    }
  }
}